// src/widgets/StudentTable/types/type.ts
export interface Student {
  id: string;
  firstName: string;
  lastName: string;
  middleName: string;
  passportNumber: string;
  jshshr: string;
  dateOfBirth: string;
  gender: string;
  nationality: string;
  email: string;
  password: string;
  phoneNumber: string;
  parentPhoneNumber: string;
}